backend-api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.1
aiomqtt==2.0.0
python-multipart==0.0.6
httpx==0.26.0
firebase-admin==6.4.0
email-validator==2.1.0
bcrypt==4.1.3
